requirements.txt
flask>=2.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
google-cloud-storage>=2.14.0
google-auth>=2.25.0
itsdangerous>=2.1.0
python-json-logger>=2.0.0
Pillow>=10.0.0
